Move route url decoding to our custom route class instead of reopening
